update check_mk version check, fix compile issue in mk-livestatus

+++ build_check_mk.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-VERSION="1.4.0p19"
+VERSION="1.4.0p21"
 SNAP7_VERSION="1.3.0"
 
 if [ $# -gt 0 ]; then
@@ -21,6 +21,11 @@ patch -p0 < ../Makefile.patch
 cp ../libssl_compat.h packages/nrpe/libssl_compat.h
 patch -p0 < ../nrpe-Makefile.patch
 
+# prepare mk-livestatus (replace usage of std::atomic_uint32_t)
+tar -xvzf packages/mk-livestatus/mk-livestatus-${VERSION}.tar.gz -C packages/mk-livestatus
+find packages/mk-livestatus/mk-livestatus-${VERSION}/src -type f -exec sed -i 's/std::atomic_int32_t/std::atomic<std::int32_t>/' "{}" +;
+tar czf packages/mk-livestatus/mk-livestatus-${VERSION}.tar.gz -C packages/mk-livestatus mk-livestatus-${VERSION}
+
 # prepare snap7
 tar -xvzf packages/snap7/snap7-full-${SNAP7_VERSION}.tar.gz -C packages/snap7
 cp packages/snap7/snap7-full-${SNAP7_VERSION}/build/unix/arm_v6_linux.mk packages/snap7/snap7-full-${SNAP7_VERSION}/build/unix/armv6l_linux.mk
@@ -31,5 +36,7 @@ tar czf packages/snap7/snap7-full-${SNAP7_VERSION}.tar.gz -C packages/snap7 snap
 make deb EDITION=raw DEBFULLNAME="Christian Hofer" DEBEMAIL=chrisss404@gmail.com
 
 # cleanup
-cd ..
-rm -rf check-mk-raw-${VERSION}.cre
+if [ $? -eq 0 ]; then
+    cd ..
+    rm -rf check-mk-raw-${VERSION}.cre
+fi
